requirements.txt
viam-sdk==0.39.0
dt-apriltags
numpy
opencv-python
pillow
typing-extensions
